准备写PinnedSectionListAdapter

+++ app/src/main/java/com/example/administrator/mytimelogger/model/Tag.java
@@ -8,18 +8,18 @@ public class Tag {
     private int id;
     private String name;
     private long color;
-    private long icon;
+    private int icon;
 
     public Tag() {
     }
 
-    public Tag(String name, long color, long icon) {
+    public Tag(String name, long color, int icon) {
         this.name = name;
         this.color = color;
         this.icon = icon;
     }
 
-    public Tag(int id, String name, long color, long icon) {
+    public Tag(int id, String name, long color, int icon) {
         this(name, color, icon);
         this.id = id;
     }
@@ -36,7 +36,7 @@ public class Tag {
         return color;
     }
 
-    public long getIcon() {
+    public int getIcon() {
         return icon;
     }
 
@@ -48,7 +48,7 @@ public class Tag {
         this.name = name;
     }
 
-    public void setIcon(long icon) {
+    public void setIcon(int icon) {
         this.icon = icon;
     }
